pyproject.toml
[tool.poetry]
name = "paylead-case-study"
version = "0.1.0"
description = ""
authors = ["Valentin Mouret <valentin.mouret@hey.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "^2.0.2"
# Used by the transaction generator to create make values.
Faker = "^18.10.1"
pyyaml = "^6.0"
clickhouse-connect = "^0.6.1"
# Library to serde JSON data.
orjson = "^3.9.0"
pydantic = "^1.10.8"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
ruff = "^0.0.270"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
